gstream:nufbdev: Removed misleading error messages
During the process of searching for the VC8000 video device, nufbdev
attempts to identify whether a video device is VC8000. If it is not,
it prints an error message that can be misleading, making it seem
like an error has occurred when it is simply not a match. nufbdev
will continue to search for other video devices.

This patch removes the misleading error message to prevent confusion.

Change-Id: Ia0ea18ea0de5ba3235207b3d017c886f1bbdbf60